clean up city fair list query renderer
