+++ src/lib/AppRegistry.tsx
@@ -11,11 +11,12 @@ import { GeneRenderer } from "./Containers/Gene"
 import { InboxRenderer } from "./Containers/Inbox"
 import { InquiryRenderer } from "./Containers/Inquiry"
 import { RegistrationFlowRenderer } from "./Containers/RegistrationFlow"
-import { CityFairListRenderer, CitySavedListRenderer, CitySectionListRenderer } from "./relay/QueryRenderers"
+import { CitySavedListRenderer, CitySectionListRenderer } from "./relay/QueryRenderers"
 import { ArtworkQueryRenderer } from "./Scenes/Artwork/Artwork"
 import { ArtworkAttributionClassFAQRenderer } from "./Scenes/ArtworkAttributionClassFAQ"
 import { CityView } from "./Scenes/City"
 import { CityBMWListRenderer } from "./Scenes/City/CityBMWList"
+import { CityFairListRenderer } from "./Scenes/City/CityFairList"
 import { CityPicker } from "./Scenes/City/CityPicker"
 import { CollectionRenderer } from "./Scenes/Collection/Collection"
 import { CollectionFullFeaturedArtistListRenderer } from "./Scenes/Collection/Components/FullFeaturedArtistList"
@@ -172,15 +173,6 @@ const ShowMoreInfo: React.SFC<ShowMoreInfoProps> = ({ showID }) => {
   return <ShowMoreInfoRenderer showID={showID} />
 }
 
-interface CityFairListProps {
-  citySlug: string
-}
-const CityFairList: React.SFC<CityFairListProps> = ({ citySlug }) => {
-  return (
-    <CityFairListRenderer citySlug={citySlug} render={renderWithLoadProgress(Containers.CityFairList, { citySlug })} />
-  )
-}
-
 interface CitySectionListProps {
   citySlug: string
   section: BucketKey
@@ -301,7 +293,7 @@ AppRegistry.registerComponent("Map", () => MapContainer)
 AppRegistry.registerComponent("City", () => CityView)
 AppRegistry.registerComponent("CityPicker", () => CityPicker)
 AppRegistry.registerComponent("CityBMWList", () => CityBMWListRenderer)
-AppRegistry.registerComponent("CityFairList", () => CityFairList)
+AppRegistry.registerComponent("CityFairList", () => CityFairListRenderer)
 AppRegistry.registerComponent("CitySavedList", () => CitySavedList)
 AppRegistry.registerComponent("CitySectionList", () => CitySectionList)
 AppRegistry.registerComponent("Collection", () => CollectionRenderer)
