+++ src/lib/Containers/index.tsx
@@ -1,5 +1,4 @@
 import { ArtworkContainer as Artwork } from "lib/Scenes/Artwork"
-import CityFairList from "lib/Scenes/City/CityFairList"
 import CitySavedList from "lib/Scenes/City/CitySavedList"
 import CitySectionList from "lib/Scenes/City/CitySectionList"
 import { PartnerContainer as Partner } from "lib/Scenes/Partner"
@@ -12,7 +11,6 @@ export default {
   Artwork,
   CitySavedList,
   CitySectionList,
-  CityFairList,
   Fair,
   Home,
   MyProfile,
