--- src/lib/Scenes/City/__tests__/CityFairList-tests.tsx
@@ -1,5 +1,0 @@
-import CityFairList from "../CityFairList"
-
-describe(CityFairList, () => {
-  it.todo("No tests yet, it's just placeholders so far.")
-})
